requirements-test.txt
homeassistant-stubs==2024.12.0
voluptuous-stubs==0.1.1

pandas==2.3.2
pvlib==0.11.1

mypy==1.14.1
pytest==8.3.3
pytest-aiohttp==1.0.5
pytest-asyncio==0.24.0
pytest-cov==6.0.0
# https://github.com/MatthewFlamm/pytest-homeassistant-custom-component/releases
pytest-homeassistant-custom-component>=0.13.190
ruff==0.8.0
